% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{heteroPopTransmission}
\alias{heteroPopTransmission}
\title{Simulating Frequency-Biased Cultural Transmission with an Heterogenous Population.}
\usage{
heteroPopTransmission(N = 500, timesteps = 301, mu = 0.01,
  bmean = 0, bsd = 0)
}
\arguments{
\item{N}{Number of individuals.}

\item{timesteps}{Number of generations (timesteps).}

\item{mu}{Innovation rate.}

\item{bmean}{Mean of the frequency bias parameter b.}

\item{bsd}{Standard deviation of the frequency bias parameter b.}
}
\value{
A vector containing the Simpson's diversity index at each timestep/generation.
}
\description{
Simulates unbiased/frequency-biased transmission with the bias parameter of each individual randomly drawn from a normal distribution with user-defined mean and standard deviation.
}
\details{
The function simulates cultural transmission within a population of heterogenous learners, each with a different levels of frequency bias at each generation. More formally the probability of an individual $x$ copying a variant $i$ is given by the following equation.
\deqn{\pi_i=\frac{\left(\frac{m_i}{N}\right)^{1-\beta_x}}{\sum\limits_{j=1}^k \left(\frac{m_j}{N}\right)^{1-\beta_x}}(1-\mu)}
where \eqn{\beta_x} is randomly drawn from a normal distribution with mean \code{bmean} and standard deviation \code{bsd}.
}
