% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{transmission}
\alias{transmission}
\title{Simulating Unbiased and Frequency-Biased Cultural Transmission}
\usage{
transmission(N = 500, timesteps = 501, mu = 0.01, warmUp = 300,
  top = NA, bias = 0, raw = FALSE)
}
\arguments{
\item{N}{Number of individuals.}

\item{timesteps}{Number of generations (timesteps).}

\item{mu}{Innovation rate.}

\item{top}{Number indicating the most common variants considered for the turn-over rate analysis (e.g. if \code{top} is equal to 5 the turnover rate is computed as the average number of new variants introduced in the top 1, 2, 3, ,4 and 5 ranked variants at each timestep).}

\item{bias}{Frequency bias parameter. Either a vector with length equal to the number of timesteps (to allow for a time-varying frequency bias), or single value (for a fixed bias).}

\item{raw}{Logical variable indicating whether the matrix of variants frequencies are returned or not. Default is FALSE.}

\item{wamUp}{Number of generations executed to reach equilibrium conditions (discarded from analysis)}
}
\value{
A list containing: 1) a matrix of variants frequencies, with rows equal to  generations and columns to specific variants (\code{rawMatrix}, available on when the argument \code{raw} is set to \code{TRUE}.; 2) a data.frame with the observed turnover rates (column: \code{obs.z}), the expected rates according to  Bentley et al 2007 (column: \code{exp.z.Bentley}) and Evans and Giometto 2012 (column: \code{exp.z.EvansGiometto}); 3) the fitted value for the parameter \code{x}; 4) a vector with a time-series of observed Simpson's diversity index (\code{obs.div}); and 5) the expected Simpson's diversity under neutrality, equal to \eqn{ 1 - (1/(2*N*\mu + 1))} (\code{exp.div}).
}
\description{
Function for simulating unbiased/frequency-biased cultural transmission.
}
\details{
The function simulates unbiased (\code{bias}=0), confomist (\code{bias}<0), and anti-conformist (\code{bias}>0) cultural transmission with a user-defined population size and mutation rates, returning Simpson's diversity index of variants frequency, observed and expected turnover rates, and estimates of the exponent \eqn{x} following the procedure described in Acerbi and Bentley 2014 (notice that Acerbi and Bentley refer to this exponent as \eqn{b}). Frequency bias is modelled using the following formula:
\deqn{\pi_i=\frac{\left(\frac{m_i}{N}\right)^{1-\beta}}{\sum\limits_{j=1}^k \left(\frac{m_j}{N}\right)^{1-\beta}}(1-\mu)}
with \eqn{pi_i} equal to the probability of copying the variant \eqn{i}, \eqn{m_i} to the number of individuals possessing the variant \eqn{i}, \eqn{N} to the population size, \eqn{mu} to the mutation rate, and \eqn{\beta} to the frequency bias parameter (i.e. \code{bias}).
}
\references{
Acerbi, A., Bentley, A.R., 2014. Biases in cultural transmission shape the turnover of populat traits. \emph{Evolution and Human Behavior}, 35, 228–236.\cr
Bentley, R.A., Lipo, C.P., Herzog, H.A., Hahn, M.W., 2007. Regular rates of popular culture change reflect random copying. \emph{Evolution and Human Behavior}, 28, 151–158. \cr
Crema, E.R., Kandler, A., Shennan, S., 2016. Revealing patterns of cultural transmission from frequency data: equilibrium and non-equilibrium assumptions. \emph{Scientific Reports} 6, 39122. https://doi.org/10.1038/srep39122 \cr
Evans, T.S., Giometto, A., 2011. Turnover Rate of Popularity Charts in Neutral Models. arXiv: [physics:soc-ph].
}
